Simplify the execution loop and remove wrong assertions
By simplifying the execution loop, the assertions in `exec_step` are
no longer required.

The assumption that the current program element has to be complete at
the beginning of `exec_step` is wrong, because there is a program element
that can change its state between invocations of `exec_step` (`breakpoint`).