Add redirect for doc urls starting with `search/`
If the `doc` url parameter is currently set to a path starting with
`search/`, the search page is shown, but no query is passed to it
unless there is a url parameter `q`.

`?doc=search/` should not show a search page in the first place,
but because that's how the path works internally, it needs to be handled.
As a compromise, this change redirects the user to the _proper_ search
url with the given parameters.