This branch implements hoisting of functions, classes, namespaces, as well as class members (name lookup within a class), addressing issue #49. It does so by introducing a second compiler pass. Since the lookup process changed considerably and many small parts needed to be deferred to the second pass, the time at which some errors are detected has changed. In the process, the structure (or "zoo") of error messages was restructured. This affects name lookup, and in some places also syntax errors. There are additional and revised unit tests covering the new lookup rules, as well as a few aspects that were not covered before (like constant initializers of default parameters and attributes).
One new complication stems from the technique of hoisting as it is applied in
javascript: it becomes possible, in principle, to access variables before they
are initialized. In these cases, the current implementation reports a runtime
error. It may be possible to catch most of these cases at compile time, but
that's outside of the scope of this branch.

The documentation was updated to reflect the changes, namely hoisting rules
and the new and changed error messages.